+++ src/js/benefits/data.js
@@ -1,0 +1,22 @@
+export const benefits = [
+  {
+    iconName: 'expertise',
+    title: 'Expertise',
+    text: 'As a highly experienced developer, I have deep knowledge of programming and website development.',
+  },
+  {
+    iconName: 'communication',
+    title: 'Communication',
+    text: ' Understanding your needs and wants is my priority and I am always open to discussions and corrections.',
+  },
+  {
+    iconName: 'art',
+    title: 'Art',
+    text: 'Thanks to my creative nature, I am ready to accept challenges and help you bring your ideas to life. ',
+  },
+  {
+    iconName: 'execution',
+    title: 'Urgent execution',
+    text: 'I understand how important time is to you. Ready to work quickly and efficiently, without reducing the quality of work.',
+  },
+];
