+++ src/js/benefits/refs.js
@@ -1,0 +1,4 @@
+export const refs = {
+  ulEl: document.querySelector('.js-benefits-list'),
+  btnEl: document.querySelector('.benefits-btn'),
+};
